split c io (works with run-c)
